structure: move modules to new dir
