+++ .travis.yml
@@ -3,7 +3,7 @@ sudo: required
 services:
     - docker
 env:
-    - GO_DIRS=". ./classifiers ./wrappers ./types ./utils ./godpi_example"
+    - GO_DIRS=". ./modules/classifiers ./modules/wrappers ./types ./utils ./godpi_example"
 script:
     - sudo docker build -t godpi-example .
     - >
